requirements.txt
python-telegram-bot~=20.3
moviepy~=1.0.3
python-dotenv~=1.0.0